WalletApp: User service and user controller

+++ WalletApp/Presentation/DTOs/UserDto.cs
@@ -3,7 +3,7 @@
     public class UserDto
     {
         public int id { get; set; }
-        public long points { get; set; }
+        public double points { get; set; }
         public decimal balance { get; set; }
     }
 }
